Move signalhandler + backtrace generation + watchdog to own file
