+++ msbuild/domoticz.vcxproj.filters
@@ -1423,6 +1423,9 @@
     <ClInclude Include="..\main\Scheduler.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\main\SignalHandler.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\main\SQLHelper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -2188,6 +2191,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\main\Scheduler.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\main\SignalHandler.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\main\SQLHelper.cpp">
